moved widget functions to widget module
