+++ src/main.rs
@@ -7,13 +7,13 @@ mod icons;
 mod stdout;
 mod task_runners;
 mod utils;
+mod widgets;
 
+use crate::widgets::action;
+use crate::widgets::target_card;
 use args::parse_args;
 use iced::alignment::Horizontal::Left;
-use iced::widget::{
-    self, button, column, container, horizontal_space, pick_list, row, scrollable, text, tooltip,
-    Column,
-};
+use iced::widget::{self, column, horizontal_space, pick_list, row, scrollable, text, Column};
 use iced::Alignment::Center;
 use iced::Length::Fill;
 use iced::{Element, Font, Subscription, Task, Theme};
@@ -316,37 +316,6 @@ impl Editor {
 
     fn theme(&self) -> Theme {
         self.theme.clone()
-    }
-}
-
-fn target_card<'a, Message: Clone + 'a>(
-    action: Element<'a, Message>,
-    label: &'a str,
-    other_action: Element<'a, Message>,
-) -> Element<'a, Message> {
-    container(row![action, other_action, label].spacing(1))
-        .style(container::bordered_box)
-        .padding(1)
-        .into()
-}
-
-fn action<'a, Message: Clone + 'a>(
-    content: impl Into<Element<'a, Message>>,
-    label: &'a str,
-    on_press: Option<Message>,
-) -> Element<'a, Message> {
-    let action = button(container(content).center_x(10));
-
-    if let Some(on_press) = on_press {
-        tooltip(
-            action.on_press(on_press),
-            label,
-            tooltip::Position::FollowCursor,
-        )
-        .style(container::transparent)
-        .into()
-    } else {
-        action.style(button::secondary).into()
     }
 }
